app/src/main/java/com/ilhmdhn/movieapps/ui/home/movie/MovieViewModel.kt
package com.ilhmdhn.movieapps.ui.home.movie

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import androidx.paging.PagedList
import com.ilhmdhn.movieapps.data.MovieRepository
import com.ilhmdhn.movieapps.data.source.local.entity.ListMovieEntity
import com.ilhmdhn.movieapps.vo.Resource

class MovieViewModel(private val movieRepository: MovieRepository) : ViewModel() {

    fun getMovies(): LiveData<Resource<PagedList<ListMovieEntity>>> = movieRepository.getAllMovie()
}